+++ dds_web/api/schemas/project_schemas.py
@@ -6,6 +6,7 @@
 
 # Standard Library
 import os
+import re
 
 # Installed
 import botocore.client
@@ -78,9 +79,7 @@ class CreateProjectSchema(marshmallow.Schema):
     description = marshmallow.fields.String(
         required=True,
         allow_none=False,
-        validate=marshmallow.validate.And(
-            marshmallow.validate.Length(min=1), dds_web.utils.contains_disallowed_characters
-        ),
+        validate=marshmallow.validate.Length(min=1),
         error_messages={
             "required": {"message": "A project description is required."},
             "null": {"message": "A project description is required."},
@@ -89,10 +88,7 @@ class CreateProjectSchema(marshmallow.Schema):
     pi = marshmallow.fields.String(
         required=True,
         allow_none=False,
-        validate=marshmallow.validate.And(
-            marshmallow.validate.Length(min=1, max=255),
-            dds_web.utils.contains_disallowed_characters,
-        ),
+        validate=marshmallow.validate.Email(),
         error_messages={
             "required": {"message": "A principal investigator is required."},
             "null": {"message": "A principal investigator is required."},
@@ -126,6 +122,15 @@ class CreateProjectSchema(marshmallow.Schema):
             ]
         ):
             raise marshmallow.ValidationError("Missing fields!")
+
+    @marshmallow.validates("description")
+    def validate_description(self, value):
+        """Verify that description only has words, spaces and . / ,."""
+        disallowed = re.findall(r"[^(\w\s.,)]+", value)
+        if disallowed:
+            raise marshmallow.ValidationError(
+                message="The description can only contain letters, spaces, period and commas."
+            )
 
     def generate_bucketname(self, public_id, created_time):
         """Create bucket name for the given project."""