+++ dds_web/api/user.py
@@ -79,8 +79,11 @@ class AddUser(flask_restful.Resource):
         send_email = json_info.get("send_email", True)
 
         # Check if email is registered to a user
-        existing_user = user_schemas.UserSchema().load({"email": email})
-        unanswered_invite = user_schemas.UnansweredInvite().load({"email": email})
+        try:
+            existing_user = user_schemas.UserSchema().load({"email": email})
+            unanswered_invite = user_schemas.UnansweredInvite().load({"email": email})
+        except sqlalchemy.exc.OperationalError as err:
+            raise ddserr.DatabaseError(message=str(err), alt_message="Unexpected database error.")
 
         if existing_user or unanswered_invite:
             if not project:
@@ -259,9 +262,17 @@ class AddUser(flask_restful.Resource):
         if goahead:
             try:
                 db.session.commit()
-            except sqlalchemy.exc.SQLAlchemyError as sqlerr:
+            except (sqlalchemy.exc.SQLAlchemyError, sqlalchemy.exc.OperationalError) as sqlerr:
                 db.session.rollback()
-                raise ddserr.DatabaseError(message=str(sqlerr))
+                raise ddserr.DatabaseError(
+                    message=str(sqlerr),
+                    alt_message=f"Invitation failed"
+                    + (
+                        ": Database malfunction."
+                        if isinstance(sqlerr, sqlalchemy.exc.OperationalError)
+                        else "."
+                    ),
+                ) from sqlerr
 
             AddUser.compose_and_send_email_to_user(
                 userobj=new_invite, mail_type="invite", link=link
@@ -371,12 +382,22 @@ class AddUser(flask_restful.Resource):
 
         try:
             db.session.commit()
-        except (sqlalchemy.exc.SQLAlchemyError, sqlalchemy.exc.IntegrityError) as err:
+        except (
+            sqlalchemy.exc.SQLAlchemyError,
+            sqlalchemy.exc.IntegrityError,
+            sqlalchemy.exc.OperationalError,
+        ) as err:
             flask.current_app.logger.exception(err)
             db.session.rollback()
             raise ddserr.DatabaseError(
-                message="Server Error: User was not associated with the project"
-            )
+                message=str(err),
+                alt_message=f"Server Error: User was not associated with the project"
+                + (
+                    ": Database malfunction."
+                    if isinstance(err, sqlalchemy.exc.OperationalError)
+                    else "."
+                ),
+            ) from err
 
         # If project is already released and not expired, send mail to user
         send_email = send_email and project.current_status == "Available"
@@ -494,7 +515,7 @@ class DeleteUserSelf(flask_restful.Resource):
     Every user can self-delete the own account with an e-mail confirmation.
     """
 
-    @auth.login_required
+    @auth.login_required(role=["Unit Admin", "Unit Personnel", "Project Owner", "Researcher"])
     @logging_bind_request
     def delete(self):
         """Request deletion of own account."""
@@ -507,6 +528,19 @@ class DeleteUserSelf(flask_restful.Resource):
         proj_ids = None
         if current_user.role != "Super Admin":
             proj_ids = [proj.public_id for proj in current_user.projects]
+
+        if current_user.role == "Unit Admin":
+            num_admins = models.UnitUser.query.filter_by(
+                unit_id=current_user.unit.id, is_admin=True
+            ).count()
+            if num_admins <= 3:
+                raise ddserr.AccessDeniedError(
+                    message=(
+                        f"Your unit only has {num_admins} Unit Admins. "
+                        "You cannot delete your account. "
+                        "Invite a new Unit Admin first if you wish to proceed."
+                    )
+                )
 
         # Create URL safe token for invitation link
         s = itsdangerous.URLSafeTimedSerializer(flask.current_app.config["SECRET_KEY"])
@@ -533,11 +567,17 @@ class DeleteUserSelf(flask_restful.Resource):
                     "status": http.HTTPStatus.OK,
                 }
 
-        except sqlalchemy.exc.SQLAlchemyError as sqlerr:
+        except (sqlalchemy.exc.SQLAlchemyError, sqlalchemy.exc.OperationalError) as sqlerr:
             db.session.rollback()
             raise ddserr.DatabaseError(
-                message=f"Creation of self-deletion request failed due to database error: {sqlerr}",
-            )
+                message=str(sqlerr),
+                alt_message=f"Creation of self-deletion request failed"
+                + (
+                    ": Database malfunction."
+                    if isinstance(sqlerr, sqlalchemy.exc.OperationalError)
+                    else "."
+                ),
+            ) from sqlerr
 
         # Create link for deletion request email
         link = flask.url_for("auth_blueprint.confirm_self_deletion", token=token, _external=True)
@@ -606,7 +646,11 @@ class UserActivation(flask_restful.Resource):
         if "email" not in json_input:
             raise ddserr.DDSArgumentError(message="User email missing.")
 
-        user = user_schemas.UserSchema().load({"email": json_input.pop("email")})
+        try:
+            user = user_schemas.UserSchema().load({"email": json_input.pop("email")})
+        except sqlalchemy.exc.OperationalError as err:
+            raise ddserr.DatabaseError(message=str(err), alt_message="Unexpected database error.")
+
         if not user:
             raise ddserr.NoSuchUserError()
 
@@ -668,9 +712,17 @@ class UserActivation(flask_restful.Resource):
 
         try:
             db.session.commit()
-        except sqlalchemy.exc.SQLAlchemyError as err:
+        except (sqlalchemy.exc.SQLAlchemyError, sqlalchemy.exc.OperationalError) as err:
             db.session.rollback()
-            raise ddserr.DatabaseError(message=str(err))
+            raise ddserr.DatabaseError(
+                message=str(err),
+                alt_message=f"Unexpected database error"
+                + (
+                    ": Database malfunction."
+                    if isinstance(err, sqlalchemy.exc.OperationalError)
+                    else "."
+                ),
+            ) from err
         msg = (
             f"The user account {user.username} ({user_email_str}, {user.role}) "
             f" has been {action}d successfully been by {current_user.name} ({current_user.role})."
@@ -700,8 +752,23 @@ class DeleteUser(flask_restful.Resource):
     @logging_bind_request
     @handle_validation_errors
     def delete(self):
+        """Delete user or invite in the DDS."""
+        current_user = auth.current_user()
 
-        user = user_schemas.UserSchema().load(flask.request.json)
+        json_info = flask.request.json
+        if json_info:
+            is_invite = json_info.pop("is_invite", False)
+            if is_invite:
+                email = self.delete_invite(email=json_info.get("email"))
+                return {
+                    "message": ("The invite connected to email " f"'{email}' has been deleted.")
+                }
+
+        try:
+            user = user_schemas.UserSchema().load(json_info)
+        except sqlalchemy.exc.OperationalError as err:
+            raise ddserr.DatabaseError(message=str(err), alt_message="Unexpected database error.")
+
         if not user:
             raise ddserr.UserDeletionError(
                 message=(
@@ -758,13 +825,55 @@ class DeleteUser(flask_restful.Resource):
             parent_user = models.User.query.get(user.username)
             db.session.delete(parent_user)
             db.session.commit()
-        except sqlalchemy.exc.SQLAlchemyError as err:
+        except (sqlalchemy.exc.SQLAlchemyError, sqlalchemy.exc.OperationalError) as err:
             db.session.rollback()
-            raise ddserr.DatabaseError(message=str(err))
+            raise ddserr.DatabaseError(
+                message=str(err),
+                alt_message=f"Failed to delete user"
+                + (
+                    ": Database malfunction."
+                    if isinstance(err, sqlalchemy.exc.OperationalError)
+                    else "."
+                ),
+            ) from err
+
+    @staticmethod
+    def delete_invite(email):
+        current_user_role = auth.current_user().role
+        try:
+            unanswered_invite = user_schemas.UnansweredInvite().load({"email": email})
+            if unanswered_invite:
+                if current_user_role == "Super Admin" or (
+                    current_user_role == "Unit Admin"
+                    and unanswered_invite.role in ["Unit Admin", "Unit Personnel", "Researcher"]
+                ):
+                    db.session.delete(unanswered_invite)
+                    db.session.commit()
+                else:
+                    raise ddserr.AccessDeniedError(
+                        message="You do not have the correct permissions to delete this invite."
+                    )
+        except (sqlalchemy.exc.SQLAlchemyError, sqlalchemy.exc.OperationalError) as err:
+            db.session.rollback()
+            flask.current_app.logger.error(
+                "The invite connected to the email "
+                f"{email or '[no email provided]'} was not deleted."
+            )
+            raise ddserr.DatabaseError(
+                message=str(err),
+                alt_message=f"Failed to delete invite"
+                + (
+                    ": Database malfunction."
+                    if isinstance(err, sqlalchemy.exc.OperationalError)
+                    else "."
+                ),
+            ) from err
+
+        return email
 
 
 class RemoveUserAssociation(flask_restful.Resource):
-    @auth.login_required
+    @auth.login_required(role=["Unit Admin", "Unit Personnel", "Project Owner", "Researcher"])
     @logging_bind_request
     @json_required
     @handle_validation_errors
@@ -777,7 +886,10 @@ class RemoveUserAssociation(flask_restful.Resource):
             raise ddserr.DDSArgumentError(message="User email missing.")
 
         # Check if email is registered to a user
-        existing_user = user_schemas.UserSchema().load({"email": user_email})
+        try:
+            existing_user = user_schemas.UserSchema().load({"email": user_email})
+        except sqlalchemy.exc.OperationalError as err:
+            raise ddserr.DatabaseError(message=str(err), alt_message="Unexpected database error.")
 
         if not existing_user:
             raise ddserr.NoSuchUserError(
@@ -804,14 +916,22 @@ class RemoveUserAssociation(flask_restful.Resource):
 
         try:
             db.session.commit()
-        except (sqlalchemy.exc.SQLAlchemyError, sqlalchemy.exc.IntegrityError) as err:
+        except (
+            sqlalchemy.exc.SQLAlchemyError,
+            sqlalchemy.exc.IntegrityError,
+            sqlalchemy.exc.OperationalError,
+        ) as err:
             flask.current_app.logger.exception(err)
             db.session.rollback()
             raise ddserr.DatabaseError(
-                message=(
-                    "Server Error: Removing user association with the project has not succeeded"
-                )
-            )
+                message=str(err),
+                alt_message=f"Server Error: Removing user association with the project has not succeeded"
+                + (
+                    ": Database malfunction."
+                    if isinstance(err, sqlalchemy.exc.OperationalError)
+                    else "."
+                ),
+            ) from err
 
         flask.current_app.logger.debug(
             f"User {existing_user.username} no longer associated with project {project.public_id}."
@@ -878,9 +998,17 @@ class ShowUsage(flask_restful.Resource):
             unit_info = models.Unit.query.filter(
                 models.Unit.id == sqlalchemy.func.binary(current_user.unit_id)
             ).first()
-        except sqlalchemy.exc.SQLAlchemyError as err:
+        except (sqlalchemy.exc.SQLAlchemyError, sqlalchemy.exc.OperationalError) as err:
             flask.current_app.logger.exception(err)
-            raise ddserr.DatabaseError("Failed getting unit information.")
+            raise ddserr.DatabaseError(
+                message=str(err),
+                alt_message=f"Failed to get unit information."
+                + (
+                    ": Database malfunction."
+                    if isinstance(err, sqlalchemy.exc.OperationalError)
+                    else "."
+                ),
+            ) from err
 
         # Total number of GB hours and cost saved in the db for the specific unit
         total_gbhours_db = 0.0